refactor: moved mocks
